feat: setup.py

+++ install.sh
@@ -2,8 +2,14 @@
 echo "This script needs the following packages: "
 echo "alien libaio1 unixodbc"
 echo "Do you want to install these packages ?"
+read -p "Are you sure you want to install these packages? "  -r
+echo    # (optional) move to a new line
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
 sudo apt install alien libaio1 unixodbc
-cd Disk1/
+
+else
+while read -r -t 0; do read -r; done
 sudo -s
 sudo cat config > /sbin/chkconfig
 sudo chmod 755 /sbin/chkconfig
@@ -13,8 +19,16 @@ sudo cat rc2-file > /etc/rc2.d/S01shm_load
 sudo chmod 755 /etc/rc2.d/S01shm_load
 sudo ln -s /usr/bin/awk /bin/awk 
 sudo mkdir /var/lock/subsys 
-sudo touch /var/lock/subsys/listener
+sudo touch /var/lock/subsys/
+echo ""
+read -p "Do you want to install Oracle SqlPlus from the package? "  -r
+echo    # (optional) move to a new line
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+echo "\nThis might take a while\n"
 sudo dpkg --install oracle-xe_11.2.0-2_amd64.deb
-echo "This might also take a while..."
+fi
 sudo /etc/init.d/oracle-xe configure 
+echo "This might also take a while..."
 echo "Oracle SQL-Plus is done configuring"
+fi
